Fix stanford-world-no-download rights for files
Connects to sul-dlss/argo#2382

This PR makes a small change to dor-rights-auth that allows for the parser to make smart decisions about files that are world access + stanford download. It does so by first combining all file-specific accesses into a single tree. Without this step, the parser would generate new rights for each file-specific access. For a particular file, the first time through (say, the world-no-download part), the access is set properly if incompletely but the next time through it would be difficult (impossible?) to tell which bits of the prior run(s) to modify and which to leave in place. By combining file accesses together, the parser can reason about the access rights for a particular file all at the same time.

Note that an existing test was modified but I am not sure that the past test behavior was useful.